Safari-611.1.10/LayoutTests/http/tests/misc/frame-default-enc-same-domain.html

<html>
<head>
<meta http-equiv="content-type" content="text/html; charset=windows-1256">
<title>Frames and encodings</title>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<frameset rows="0,*">
    <frame src="about:" scrolling=no marginwidth=0 marginheight=0>
    <frame src="resources/frame-default-enc-frame.html">
</frameset>
</html>
