safari-534.53-branch/LayoutTests/http/tests/misc/frame-default-enc-same-domain-expected.txt

ALERT: windows-1256
